+++ src/test/java/com/openggf/game/rewind/TestScalarOnlyCodecDeletion.java
@@ -1269,6 +1269,11 @@ public class TestScalarOnlyCodecDeletion {
                             "com.openggf.game.sonic3k.objects.TensionBridgeObjectInstance",
                             "baseY", "negativeSubtype", "segmentCount", "triggerIndex"));
 
+    private static final List<MutableFieldCoverageCandidate> BATCH181_MUTABLE_FIELDS =
+            List.of(new MutableFieldCoverageCandidate(
+                    "com.openggf.game.sonic3k.objects.MGZPulleyObjectInstance",
+                    "anchorX", "anchorY", "flipped", "targetExtension"));
+
     private static final List<CodecDeletionCandidate> BATCH31_DELETED_CODECS = List.of(
             new CodecDeletionCandidate(BombPrizeObjectInstance.class.getName(), GameId.S2));
 
@@ -5248,6 +5253,23 @@ public class TestScalarOnlyCodecDeletion {
     @Test
     void batch180MechanismRootScalarsAreMutableForCompactRestore() {
         for (MutableFieldCoverageCandidate candidate : BATCH180_MUTABLE_FIELDS) {
+            Class<?> cls = loadClass(candidate.fqn());
+            for (String fieldName : candidate.fieldNames()) {
+                try {
+                    var field = findField(cls, fieldName);
+                    assertFalse(Modifier.isFinal(field.getModifiers()),
+                            cls.getName() + "#" + fieldName
+                                    + " must be mutable so compact restore can replay captured scalars");
+                } catch (NoSuchFieldException e) {
+                    throw new AssertionError("Missing scalar field " + cls.getName() + "#" + fieldName, e);
+                }
+            }
+        }
+    }
+
+    @Test
+    void batch181PulleyRootScalarsAreMutableForCompactRestore() {
+        for (MutableFieldCoverageCandidate candidate : BATCH181_MUTABLE_FIELDS) {
             Class<?> cls = loadClass(candidate.fqn());
             for (String fieldName : candidate.fieldNames()) {
                 try {